\name{check_constraints.data.frame.TableChecks}
\alias{check_constraints.data.frame.TableChecks}
\title{Validate Data Frame}
\usage{
  check_constraints.data.frame.TableChecks(x, checks)
}
\arguments{
  \item{x}{\code{data.frame}. Data frame to be checked.}

  \item{columns}{\code{ColumnList} containing column level
  checks.}
}
\description{
  Validate Data Frame
}

